chore(data-warehouse): move custom source tests into __tests__ folder
