+++ products/data_warehouse/frontend/shared/components/forms/__tests__/customSourceManifest.test.ts
@@ -1,4 +1,4 @@
-import { buildManifest, extractAuthSecrets, ManifestState, parseManifestIntoState } from './customSourceManifest'
+import { buildManifest, extractAuthSecrets, ManifestState, parseManifestIntoState } from '../customSourceManifest'
 
 const baseState = (): ManifestState => ({
     base_url: 'https://api.example.com',
